add delete list dialog and pin list

+++ lib/featrues/home/presentation/views/widgets/items_view_appbar.dart
@@ -1,4 +1,6 @@
+import 'package:final_project/core/utils/app_colors.dart';
 import 'package:final_project/featrues/home/presentation/view_model/list_cubit/list_cubit.dart';
+import 'package:final_project/featrues/home/presentation/views/widgets/custom_button.dart';
 import 'package:final_project/featrues/home/presentation/views/widgets/custom_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -33,26 +35,54 @@ class ItemsViewAppBar extends StatelessWidget {
           ),
         ),
         PopupMenuButton<String>(
-          onSelected: (value) {
+          onSelected: (value) async {
             if (value == 'rename') {
               _renameList(context);
             } else if (value == 'delete') {
-              context.read<ListCubit>().deleteList(listModel.id);
-              Navigator.pop(context);
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(content: Text('Delete selected')),
+              await showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: const Column(
+                      children: [
+                        Text(
+                          'Are you sure?',
+                          style: TextStyle(fontWeight: FontWeight.bold),
+                        ),
+                        SizedBox(height: 6),
+                        Text(
+                          'Do you really want to delete this list? You will not be able to undo this action',
+                          textAlign: TextAlign.center,
+                          style: TextStyle(fontSize: 16, color: AppColors.grey),
+                        ),
+                      ],
+                    ),
+                    actions: [
+                      CustomButton(
+                        title: 'no',
+                        onPressed: () => Navigator.pop(context),
+                      ),
+                      SizedBox(height: 6),
+                      CustomButton(
+                        title: 'yes',
+                        onPressed: () async {
+                          context.read<ListCubit>().deleteList(listModel.id);
+                          Navigator.pop(context);
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            const SnackBar(content: Text('Delete selected')),
+                          );
+                          Navigator.pop(context);
+                        },
+                      ),
+                    ],
+                  );
+                },
               );
             }
           },
           itemBuilder: (context) => const [
-            PopupMenuItem(
-              value: 'rename',
-              child: Text('Rename'),
-            ),
-            PopupMenuItem(
-              value: 'delete',
-              child: Text('Delete'),
-            ),
+            PopupMenuItem(value: 'rename', child: Text('Rename')),
+            PopupMenuItem(value: 'delete', child: Text('Delete')),
           ],
         ),
       ],
@@ -89,13 +119,12 @@ class ItemsViewAppBar extends StatelessWidget {
                 );
                 onRename(newName);
 
-                if (context.mounted) Navigator.pop(context);
-
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('List renamed successfully'),
-                  ),
-                );
+                if (context.mounted) {
+                  Navigator.pop(context);
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text('List renamed successfully')),
+                  );
+                }
               }
             },
             child: const Text("Save"),
